MMIOT-232: Solve an issue to write lock state
When writing the lock state, write_lock_state() returns
TEE_ERROR_ACCESS_CONFLICT. This error occurs because
TEE_CreatePersistentObject() tries to overwrite the file
but the file was already opened by TEE_OpenPersistentObject().
Use TEE_SeekObjectData() and TEE_WriteObjectData() instead.

Note that, when the lock state is updated, the rollback
indexes are preserved.

Signed-off-by: Alexandre Jutras <alexandre.jutras@nxp.com>